+++ src/lib.rs
@@ -105,6 +105,7 @@ pub(crate) mod handles;
 pub mod logging;
 pub(crate) mod panic;
 pub(crate) mod param_convert;
+pub mod prompt;
 pub(crate) mod query_timer;
 mod sync;
 pub mod synthetic;